+++ src/components/Forms/SearchForm/MoreFilters/index.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { IoIosArrowDown } from "react-icons/io";
 import { IoIosArrowUp } from "react-icons/io";
-import { useSearchStore } from "../../../stores/useSearchStore.js";
-import Checkbox from "../../Inputs/Checkbox";
+import { useSearchStore } from "../../../../stores/useSearchStore.js";
+import Checkbox from "../../../Inputs/Checkbox/index.jsx";
 
 export default function MoreFilters({ color, mainSearch }) {
   const [openMoreFilters, setOpenMoreFilters] = useState(false);
